Add more tests, add deploy script

+++ deploy_pypi.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
-rm -rf dist
-python setup.py sdist
-cd dist
-twine upload -r pypi *.tar.gz
-cd ..
+python3 TestRoundCreator/TestMain.py
+if [ "$?" = "0" ]; then
+	rm -rf dist
+	python setup.py sdist
+	cd dist
+	twine upload -r pypi *.tar.gz
+	cd ..
+else
+  echo "Skipping deployment"
+fi
